ticket-classification-system/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
kafka-python==2.0.2
scikit-learn==1.3.2
